# Reuse resetLevels to validateDAG of segmented fusion.